Removing/Renaming toolstore references & addressing comments
1.Removal of any references to the toolstore.
2. Addressing comments regarding variable naming & order of operations.
