+++ app/src/main/java/com/example/vu/android/empowerplant/StoreItem.java
@@ -1,4 +1,4 @@
-package com.example.vu.android.toolstore;
+package com.example.vu.android.empowerplant;
 
 import androidx.fragment.app.Fragment;
