+++ app/src/main/res/layout/activity_empowerplant.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".toolstore.ToolStoreActivity">
+    tools:context=".empowerplant.EmpowerPlantActivity">
 
     <LinearLayout android:id="@+id/main_list"
         android:layout_width="match_parent"
